Refactor: Move helper functions to utils file and redirect user
